Reorganize hs-get-capabilities-error component
Fixes "no factory for HsGetCapabilitiesErrorComponent found" error
