+++ components/add-layers/common/capabilities-error-dialog.component.ts
@@ -1,10 +1,10 @@
 import {Component, Input, ViewRef} from '@angular/core';
-import {HsDialogComponent} from '../layout/dialogs/dialog-component.interface';
-import {HsDialogContainerService} from '../layout/dialogs/dialog-container.service';
+import {HsDialogComponent} from '../../layout/dialogs/dialog-component.interface';
+import {HsDialogContainerService} from '../../layout/dialogs/dialog-container.service';
 
 @Component({
   selector: 'hs-get-capabilities-error',
-  template: require('./partials/dialog_getcapabilities_error.html'),
+  template: require('./capabilities-error-dialog.html'),
 })
 export class HsGetCapabilitiesErrorComponent implements HsDialogComponent {
   @Input() data: any;
