+++ components/add-layers/common/capabilities-error-dialog.html
