+++ components/add-layers/wfs/index.ts
@@ -4,7 +4,7 @@ import * as angular from 'angular';
 import {HsAddLayersWfsComponent} from './add-layers-wfs.component';
 import {HsAddLayersWfsModule} from './add-layers-wfs.module';
 import {HsAddLayersWfsService} from './add-layers-wfs.service';
-import {HsGetCapabilitiesErrorComponent} from '../capabilities-error.component';
+import {HsGetCapabilitiesErrorComponent} from '../common/capabilities-error-dialog.component';
 import {downgradeComponent, downgradeInjectable} from '@angular/upgrade/static';
 
 import {downgrade} from '../../../common/downgrader';
