+++ components/add-layers/wms/index.ts
@@ -5,7 +5,7 @@ import '../../utils';
 import {HsAddLayersWmsComponent} from './add-layers-wms.component';
 import {HsAddLayersWmsModule} from './add-layers-wms.module';
 import {HsAddLayersWmsService} from './add-layers-wms.service';
-import {HsGetCapabilitiesErrorComponent} from '../capabilities-error.component';
+import {HsGetCapabilitiesErrorComponent} from '../common/capabilities-error-dialog.component';
 import {HsResampleDialogComponent} from '../resample-dialog.component';
 import {downgrade} from '../../../common/downgrader';
 import {downgradeComponent, downgradeInjectable} from '@angular/upgrade/static';
